added middleware

+++ controllers/blogs.js
@@ -4,19 +4,10 @@ const Blog = require('../models/blog');
 const User = require('../models/user');
 const jwt = require('jsonwebtoken');
 
-const getTokenFrom = (request) => {
-    const authorization = request.get('authorization');
-    if (authorization && authorization.toLowerCase().startsWith('bearer ')) {
-        return authorization.substring(7);
-    }
-    return null;
-};
-
 blogsRouter.get('/', async (request, response) => {
     const blogs = await Blog.find({}).populate('user', {
         username: 1,
-        name: 1,
-        author
+        name: 1
     });
 
     response.json(blogs);
@@ -33,15 +24,7 @@ blogsRouter.get('/:id', async (request, response) => {
 
 blogsRouter.post('/', async (request, response) => {
     const body = request.body;
-    const token = getTokenFrom(request);
-    const decodedToken = jwt.verify(token, process.env.SECRET);
-    console.log(decodedToken);
-
-    if (!token || !decodedToken.id) {
-        return response.status(401).json({ error: 'token missing or invalid' });
-    }
-
-    const user = await User.findById(decodedToken.id);
+    const user = request.user;
 
     const blog = new Blog({
         url: body.url,
@@ -58,8 +41,15 @@ blogsRouter.post('/', async (request, response) => {
 });
 
 blogsRouter.delete('/:id', async (request, response) => {
-    await Blog.findByIdAndRemove(request.params.id);
-    response.status(204).end();
+    const user = request.user;
+    const blog = await Blog.findById(request.params.id);
+
+    if (blog.user.toString() === user.id.toString()) {
+        await Blog.findByIdAndRemove(request.params.id);
+        response.status(204).end();
+    } else {
+        return response.status(401).json({ error: 'Only the creator of the blog can delete it' });
+    }
 });
 
 blogsRouter.put('/:id', async (request, response) => {
